Combine tests that call the butler into one class.
