+++ tests/testDcrTemplate.py
@@ -26,16 +26,28 @@ import numpy as np
 import unittest
 
 from lsst.afw.geom import Angle
+import lsst.daf.persistence as daf_persistence
 import lsst.utils.tests
 
 from python.dcr_utils import wrap_warpExposure
 from python.dcr_utils import calculate_rotation_angle
+from python.generateTemplate import GenerateTemplate
 from python.test_utils import BasicGenerateTemplate
 from python.test_utils import DcrModelTestBase
 
 
 class DcrTemplateTestCase(DcrModelTestBase, lsst.utils.tests.TestCase):
     """Tests for the functions in the GenerateTemplate class."""
+
+    @classmethod
+    def setUpClass(self):
+        """Set up one instance of the butler for all tests of persistence."""
+        self.repository = "./test_data/"
+        self.butler = daf_persistence.Butler(inputs=self.repository)
+
+    @classmethod
+    def tearDownClass(self):
+        del self.butler
 
     def test_dataId_single(self):
         """Test that the dataId for the `calexp` data type is correct for a single observation."""
@@ -62,9 +74,21 @@ class DcrTemplateTestCase(DcrModelTestBase, lsst.utils.tests.TestCase):
         dataId = BasicGenerateTemplate._build_model_dataId(band_ref, subfilter=subfilter)
         self.assertEqual(ref_id, dataId)
 
+    def test_simple_phase_kernel(self):
+        """Compare the result of _calc_offset_phase to previously computed values."""
+        data_file = "test_data/simple_phase_kernel.npy"
+        psf = self.exposure.getPsf()
+        psf_size = psf.computeKernelImage().getArray().shape[0]
+        phase_arr = BasicGenerateTemplate._calc_offset_phase(exposure=self.exposure,
+                                                             dcr_gen=self.dcr_gen, size=psf_size)
+        # Uncomment the following code to over-write the reference data:
+        # np.save(data_file, phase_arr, allow_pickle=False)
+        phase_arr_ref = np.load(data_file)
+        self.assertFloatsAlmostEqual(phase_arr, phase_arr_ref)
+
     def test_generate_template(self):
         """Compare the result of generate_templates_from_model to previously computed values."""
-        repository = "./test_data/"
+        self.dcrTemplate.butler = self.butler
         elevation_arr = np.radians([50., 70., 85.])
         az = Angle(0.)
         # Note that self.array is randomly generated each call. That's okay, because the template should
@@ -75,10 +99,10 @@ class DcrTemplateTestCase(DcrModelTestBase, lsst.utils.tests.TestCase):
             exposures.append(self.dcrTemplate.create_exposure(self.array, variance=None, exposureId=obsid,
                                                               elevation=Angle(el), azimuth=az))
         template_gen = self.dcrTemplate.generate_templates_from_model(exposures=exposures)
-        # Uncomment the following code to re-generate the reference data:
+        # Uncomment the following code to over-write the reference data:
         # for exposure in model_gen:
-        #     self.dcrTemplate.write_exposure(exposure, output_repository=repository)
-        template_ref_gen = self.dcrTemplate.read_exposures(obsids=obsids, input_repository=repository)
+        #     self.dcrTemplate.write_exposure(exposure, output_repository=self.repository)
+        template_ref_gen = self.dcrTemplate.read_exposures(obsids=obsids)
 
         for template_test, template_ref in izip(template_gen, template_ref_gen):
             self.assertMaskedImagesNearlyEqual(template_test.getMaskedImage(), template_ref.getMaskedImage())
@@ -99,6 +123,63 @@ class DcrTemplateTestCase(DcrModelTestBase, lsst.utils.tests.TestCase):
         """Test that we can calculate the same rotation angle that was originally supplied in setup."""
         rotation_angle = calculate_rotation_angle(self.exposure)
         self.assertAnglesNearlyEqual(self.rotation_angle, rotation_angle)
+
+    def test_create_exposure(self):
+        """Test that the data retrieved from an exposure is the same as what it was initialized with."""
+        self.assertFloatsAlmostEqual(self.exposure.getMaskedImage().getImage().getArray(), self.array)
+
+        # Check that the required metadata is present:
+        visitInfo = self.exposure.getInfo().getVisitInfo()
+        el = visitInfo.getBoresightAzAlt().getLatitude()
+        az = visitInfo.getBoresightAzAlt().getLongitude()
+        self.assertAnglesNearlyEqual(el, self.elevation)
+        self.assertAnglesNearlyEqual(az, self.azimuth)
+        hour_angle = visitInfo.getBoresightHourAngle()
+        self.assertAnglesNearlyEqual(hour_angle, self.hour_angle)
+
+    def test_persist_dcr_model_roundtrip(self):
+        """Test that an exposure can be persisted and later depersisted from a repository."""
+        self.dcrTemplate.butler = self.butler
+        self.dcrTemplate.export_model()
+
+        # First test that the model values are not changed from what is expected
+        # This requires the full GenerateTemplate class, not just the lightweight test class.
+        dcrTemplate2 = GenerateTemplate(butler=self.butler)
+        # Note that butler.get() reads the FITS file in 32 bit precision.
+        for m_new, m_ref in izip(dcrTemplate2.model, self.dcrTemplate.model):
+            self.assertFloatsAlmostEqual(m_new, m_ref, rtol=1e-7)
+
+        # Next, test that the required parameters have been restored
+        param_ref = self.dcrTemplate.__dict__
+        param_new = dcrTemplate2.__dict__
+        for key in param_ref:
+            self.assertIn(key, param_new)
+        # If the parameters are present, now check that they have the correct values.
+        # Note that this only tests floats, np.ndarrays, and strings.
+        for key in param_ref:
+            val_new = param_new.get(key)
+            val_ref = param_ref.get(key)
+            valid_float = False
+            valid_string = False
+            # Check whether the key value is a type we can test with assertFloatsAlmostEqual
+            # by testing the reference value against itself.
+            try:
+                self.assertFloatsAlmostEqual(val_ref, val_ref)
+                valid_float = True
+            except:
+                if isinstance(val_ref, str):
+                    valid_string = True
+            if valid_float:
+                print("Checking value of key: %s" % key)
+                self.assertFloatsAlmostEqual(val_new, val_ref)
+            elif valid_string:
+                print("Checking value of key: %s" % key)
+                try:
+                    self.assertEqual(val_new, val_ref)
+                except:
+                    print("Failed for some reason!")
+            else:
+                print("Skipping key: %s" % key)
 
 
 class MemoryTester(lsst.utils.tests.MemoryTestCase):
